catalog,catalog-react,techdocs: extract experimental entity page implementation from app-next
